+++ src/managers/WindowResizeManager.tsx
@@ -20,5 +20,5 @@ export function WindowResizeManager() {
     };
   }, []);
 
-  return null;
+  return <></>;
 }